Cleanups and organisation of d8 and r8
